% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-simplify.R
\name{sparseAssay}
\alias{sparseAssay}
\alias{denseAssay}
\alias{disjointAssay}
\title{Create simplified representation of ragged assay data.}
\usage{
sparseAssay(x, i = 1, withDimnames = TRUE)

denseAssay(x, i = 1, withDimnames = TRUE)

disjointAssay(x, simplify, i = 1, withDimnames = TRUE)
}
\arguments{
\item{x}{\code{RaggedExperiment}}

\item{i}{integer(1) or character(1) name of assay to be
transformed.}

\item{withDimnames}{logical(1) include dimnames on the returned
matrix. When there are no explict rownames, these are
manufactured with \code{as.character(rowRanges(x))}; rownames
are always manufactured for \code{denseAssay()} and
\code{disjointAssay()}.}

\item{simplify}{A function operating on a \code{*List}, where the
    elements of the list are all within-sample assay values from
    ranges overlapping each disjoint range. For instance, to use
    the \code{simplify=mean} of overlapping ranges, where ranges
    are characterized by integer-valued scores, the entries are
    calculated as \preformatted{
                    a
    original: |-----------|
                        b
                   |----------|

                a    a, b   b
    disjoint: |----|------|---|

    values <- IntegerList(a, c(a, b), b)
    simplify(values)
    }}
}
\value{
\code{sparseAssay()} returns a matrix() with dimensions
    \code{dim(x)}. Elements contain the assay value for the ith
    range and jth sample.

\code{denseAssay()}: Samples with identical range are placed
    in the same row. Non-disjoint ranges are NOT collapsed.

\code{disjointAssay()}: A matrix with number of rows equal
    to number of disjoint ranges across all samples. Elements of
    the matrix are summarized by applying \code{simplify()} to
    assay values of overlapping ranges
}
\description{
These methods transform \code{assay()} from the
    default (i.e., \code{sparseAssay()}) representation to various
    forms of more dense representation. \code{denseAssay()}
    collapses identical ranges across samples into a single
    row. \code{disjoinAssay()} creates disjoint (non-overlapping)
    regions, simplifies values within each sample in a
    user-specified manner, and returns a matrix of disjoint regions
    x samples.
}
