% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-functions.R
\name{sparseSummarizedExperiment}
\alias{sparseSummarizedExperiment}
\alias{compactSummarizedExperiment}
\alias{disjoinSummarizedExperiment}
\alias{qreduceSummarizedExperiment}
\title{Create SummarizedExperiment representations by transforming
    ragged assays to rectangular form.}
\usage{
sparseSummarizedExperiment(x, i = 1)

compactSummarizedExperiment(x, i = 1)

disjoinSummarizedExperiment(x, simplify, i = 1L)

qreduceSummarizedExperiment(x, query, simplify, i = 1L)
}
\arguments{
\item{x}{\code{RaggedExperiment}}

\item{i}{\code{integer(1)} or \code{character(1)} selecting the
assay to be transformed.}

\item{simplify}{\code{function} of 1 (for
\code{disjoinSummarizedExperiment}) or 3 (for
\code{qreduceSummarizedExperiment}) arguments, used to
transform assays. See \code{\link{assay-functions}}.}

\item{query}{\code{GRanges} provding regions over which reduction
is to occur.}
}
\value{
All functions return \code{RangedSummarizedExperiment}.

\code{sparseSummarizedExperiment} has \code{rowRanges()}
    identical to the row ranges of \code{x}, and \code{assay()}
    data as \code{sparseAssay()}. This is very space-inefficient
    representation of ragged data.

\code{compactSummarizedExperiment} has \code{rowRanges()}
    identical to the row ranges of \code{x}, and \code{assay()}
    data as \code{compactAssay()}. This is space-inefficient
    representation of ragged data when samples are primarily
    composed of different ranges.

\code{disjoinSummarizedExperiment} has \code{rowRanges()}
    identical to the disjoint row ranges of \code{x},
    \code{disjoint(rowRanges(x))}, and \code{assay()} data as
    \code{disjoinAssay()}.

\code{qreduceSummarizedExperiment} has \code{rowRanges()}
    identical to \code{query}, and \code{assay()} data as
    \code{qreduceAssay()}.
}
\description{
These methods transform \code{RaggedExperiment}
    objects to similar \code{SummarizedExperiment} objects. They do
    so by transforming assay data to more rectangular
    representations, following the rules outlined for similarly
    names transformations \code{sparseAssay()},
    \code{compactAssay()}, \code{disjoinAssay()}, and
    \code{qreduceAssay()}. Because of the complexity of the
    transformation, ti usually only makes sense transform
    \code{RaggedExperiment} objects with a single assay; this is
    currently enforced at time of coercion.
}
\examples{
x <- RaggedExperiment(GRangesList(
    GRanges(c("A:1-5", "A:4-6", "A:10-15"), score=1:3),
    GRanges(c("A:1-5", "B:1-3"), score=4:5)
))

## sparseSummarizedExperiment

sse <- sparseSummarizedExperiment(x)
assay(sse)
rowRanges(sse)

## compactSummarizedExperiment

cse <- compactSummarizedExperiment(x)
assay(cse)
rowRanges(cse)

## disjoinSummarizedExperiment

disjoinAssay(x, lengths)
dse <- disjoinSummarizedExperiment(x, lengths)
assay(dse)
rowRanges(dse)

## qreduceSummarizedExperiment

x <- RaggedExperiment(GRangesList(
    GRanges(c("A:1-3", "A:4-5", "A:10-15"), score=1:3),
    GRanges(c("A:4-5", "B:1-3"), score=4:5)
))
query <- GRanges(c("A:1-2", "A:4-5", "B:1-5"))

weightedmean <- function(scores, ranges, qranges)
    ## weighted average score per query range
    sum(scores * width(ranges)) / sum(width(ranges))

qreduceAssay(x, query, weightedmean)
qse <- qreduceSummarizedExperiment(x, query, weightedmean)
assay(qse)
rowRanges(qse)
}
